+++ public/browser/chunk-CHhn8gAg2.js
@@ -1,1 +1,0 @@
-import"./chunk-CD52DvrH.js";import"./chunk-B_m5Yd57.js";import"./chunk-wEifUrYN.js";import"./chunk-DIXbp78F.js";import"./chunk-Be5y8EfF.js";import"./chunk-BgjWCWRg.js";import"./chunk-1kTwg8yH.js";import"./chunk-BZd4xebm.js";import"./chunk-Ck4Fw73x.js";import{M as Fe}from"./main-6RGDWVDX.js";export{Fe as BpmnEditorPage};