+++ public/browser/chunk-DrEy4ZeB.js
@@ -1,0 +1,1 @@
+import"./chunk-CD52DvrH.js";var t=[{path:``,loadComponent:()=>import(`./chunk-CxGcO13L.js`).then(o=>o.DefinitionsListPageComponent)}];export{t as DEFINITION_ROUTES};