+++ public/browser/chunk-qWRibJ2-.js
@@ -1,0 +1,1 @@
+import"./chunk-CD52DvrH.js";var t=[{path:``,loadComponent:()=>import(`./chunk-DXaiOlQJ2.js`).then(o=>o.CallRecordsListComponent)},{path:`:id`,loadComponent:()=>import(`./chunk-C1pwskOi.js`).then(o=>o.CallRecordDetailComponent)}];export{t as CALL_ROUTES};